python/examples/requirements.txt
numpy==1.26
scipy==1.14
meshio==5.3.5
libigl==v2.5.1
polyscope==2.5.0
ilupp==1.0.2
ipctk==1.2.0
qpsolvers[open_source_solvers]==4.4.0